feat(EncodingConverter): add support for .нбр files
Added support for .нбр files to the list of file extensions that can be converted by the EncodingConverter. This will allow users to convert files in this format to the target encoding specified in the settings.

Also made some improvements to the EncodingConverter class, including adding a function to check if a string is in UTF-8 format and optimizing the iconvConvert function.

Finally, added an optimization flag to the CMakeLists file to improve performance.

Time spent: 2 hours.